Fixed database bug, added some CSS styling
Fixed a database bug that occured if a user had any '.' characters in
their email address, since SQL uses this character as a schema extender
character, everything before the '.' would be added as a new schema and
everything after was the table name. Also, added some CSS styling and
added a "Register" button to the login page.

+++ FixedIt/src/fixedIt/modelComponents/Authenticator.java
@@ -194,7 +194,7 @@ public class Authenticator implements EmailSender {
 						c.setCRN(Integer.parseInt(course.getString("crn")));					//
 						c.setCourseAndSection(course.getString("courseandsection"));			//
 						c.setTitle(course.getString("title"));									//
-						c.setCredits(Double.parseDouble(course.getString("credits")));			//
+						c.setCredits(Float.parseFloat(course.getString("credits")));			//
 						c.setType(course.getString("type"));									//
 						c.setDays(course.getString("days"));									//	add all course info
 						c.setTime(course.getString("time"));									//	to a course object
